+++ src/Products/PloneMeeting/browser/overrides.py
@@ -348,8 +348,6 @@ class PMDocumentGeneratorLinksViewlet(DocumentGeneratorLinksViewlet, BaseGenerat
 class PMDashboardDocumentGeneratorLinksViewlet(DashboardDocumentGeneratorLinksViewlet, BaseGeneratorLinksViewlet):
     """ """
 
-    render = ViewPageTemplateFile('templates/generationlinks.pt')
-
     def available(self):
         """
             Do not display it on Meeting faceted (available/presented items)