+++ src/Products/PloneMeeting/migrations/migrate_to_4108.py
@@ -22,6 +22,7 @@ class Migrate_To_4108(Migrator):
         logger.info('Done.')
 
     def fix_email_from_address(self):
+        logger.info("Fixing email from address...")
         mail_panel_adapter = MailControlPanelAdapter(self.portal)
         mail_address = mail_panel_adapter.get_email_from_address().strip()
         if mail_address.endswith("@imio.be"):
@@ -29,6 +30,7 @@ class Migrate_To_4108(Migrator):
             mail_address = public_url.replace("https://", "")\
                 .replace("-pm.imio-app", "-delib@imio")
         mail_panel_adapter.set_email_from_address(mail_address.strip())
+        logger.info('Done.')
 
     def run(self, from_migration_to_41=False):
         logger.info('Migrating to PloneMeeting 4108...')
@@ -37,6 +39,8 @@ class Migrate_To_4108(Migrator):
         # that used omittedSuffixes instead omitted_suffixes
         self.updateTALConditions(old_word='omittedSuffixes', new_word='omitted_suffixes')
         self.fix_email_from_address()
+        # reapply typeinfo as directory schema_policy changed
+        self.ps.runImportStepFromProfile('profile-Products.PloneMeeting:default', 'typeinfo')
 
 
 def migrate(context):
@@ -44,7 +48,8 @@ def migrate(context):
 
        1) Make sure format of DashboardCollection.query is correct;
        2) Fix condition for 'searchmyitemstakenover'.
-       3) Fix mail sender address.
+       3) Fix mail sender address;
+       4) Re-apply typeinfo step to update directory schema policy.
     '''
     migrator = Migrate_To_4108(context)
     migrator.run()